requirements.txt
Django==3.2.9
django_otp==1.1.1
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
django-cors-headers==3.10.0
